img2text/requirements.txt
Flask==3.0.0
pytesseract>=0.3.13
Pillow>=10.4.0
pdf2image==1.16.3
Werkzeug==3.0.1
pandas>=2.0.0
openpyxl>=3.1.0
pyinstaller>=6.0.0
